retail_research_agent/requirements.txt
# Supported: Python 3.11, 3.12, 3.13 only.
# Python 3.14: CrewAI 1.x declares Requires-Python <3.14; CrewAI 0.11.x pins ancient LangChain
# incompatible with this project. Use Python 3.13 for the venv (see README).
tiktoken>=0.12.0

# [litellm] enables Groq, Ollama, and other providers CrewAI routes via LiteLLM
crewai[litellm]>=1.10.0,<1.12.0

langchain-core>=0.3.26,<0.4.0
langchain-openai>=0.2.14,<0.4.0
langchain-community>=0.3.16,<0.4.0

tavily-python==0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
aiohttp>=3.9.0,<4.0.0

faiss-cpu>=1.12.0,<2.0.0

sentence-transformers>=2.2.0,<6.0.0

pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0.0

tenacity>=8.2,<10.0
pytest>=8.0,<9.0

streamlit>=1.33.0,<2.0.0
